chore: 주문하기 suspense 사용하여 리팩터링
