+++ src/pages/OrderPage.tsx
@@ -1,27 +1,11 @@
-import { useState, useEffect } from 'react';
-import { useParams } from 'react-router-dom';
-import { useForm } from 'react-hook-form';
 import styled from '@emotion/styled';
-import { toast } from 'react-toastify';
+import { Suspense } from 'react';
+import Spinner from '@/components/common/Spinner';
+import { ErrorBoundary } from '@/ErrorBoundary';
 
 import MobileLayout from '@/layouts/MobileLayout';
 import NavBar from '@/components/NavBar';
-import CardPicker from '@/components/order/CardPicker';
-import CardMessage from '@/components/order/CardMessage';
-import SenderInfo from '@/components/order/SenderInfo';
-import ReceiverInfo from '@/components/order/ReceiverInfo';
-import ProductInfo from '@/components/order/ProductInfo';
-import OrderButton from '@/components/order/OrderButton';
-
-import { useProductSummary } from '@/hooks/useProductSummary';
-import { useCreateOrder } from '@/hooks/useCreateOrder';
-import { useAuth } from '@/hooks/useAuth';
-import { useGoToHome } from '@/hooks/useGoTo';
-import { useGoToLogin } from '@/hooks/useGoTo';
-import { isBlank } from '@/utils/validation';
-import { cardTemplates } from '@/mock/cardTemplates';
-import type { CardTemplate } from '@/mock/cardTemplates';
-import type { Receiver } from '@/types/order';
+import OrderContent from '@/components/order/OrderContent';
 
 const Wrapper = styled.div`
   display: flex;
@@ -30,137 +14,30 @@ const Wrapper = styled.div`
   background: ${({ theme }) => theme.colors.gray[200]};
 `;
 
-interface FormValues {
-  message: string;
-  sender: string;
-  name: string;
-  phone: string;
-  qty: number;
-}
+const SpinnerWrapper = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100dvh;
+  background-color: #fff;
+`;
 
 export default function OrderPage() {
-  const goToHome = useGoToHome();
-  const goToLogin = useGoToLogin();
-  const { user } = useAuth();
-  const { id } = useParams<{ id: string }>();
-  const { product, loading } = useProductSummary(id);
-  const { mutateAsync } = useCreateOrder();
-  const defaultTpl = cardTemplates[0];
-  const [selectedTemplate, setSelectedTemplate] = useState<CardTemplate>(defaultTpl);
-  const [receivers, setReceivers] = useState<Receiver[]>([]);
-
-  const {
-    register,
-    handleSubmit,
-    setValue,
-    formState: { errors },
-  } = useForm<FormValues>({
-    defaultValues: {
-      message: defaultTpl.defaultTextMessage,
-      sender: '',
-      name: '',
-      phone: '',
-      qty: 1,
-    },
-  });
-
-  useEffect(() => {
-    if (user?.name) {
-      setValue('sender', user.name);
-    }
-  }, [user, setValue]);
-
-  if (loading) return <div>로딩 중...</div>;
-  if (!product) return <div>상품을 찾을 수 없습니다.</div>;
-
-  const onSubmit = async (data: FormValues) => {
-    if (receivers.length === 0) {
-      toast.error('받는 사람을 한 명 이상 등록해주세요.');
-      return;
-    }
-
-    const payload = {
-      productId: product.id,
-      message: data.message,
-      messageCardId: defaultTpl.id.toString(),
-      ordererName: data.sender,
-      receivers: receivers.map((r) => ({
-        name: r.name,
-        phoneNumber: r.phone,
-        quantity: r.qty,
-      })),
-    };
-
-    try {
-      await mutateAsync({
-        payload,
-        token: user?.authToken ?? '',
-      });
-
-      toast.success(
-        <div>
-          주문이 완료되었습니다.
-          <br />
-          상품명: {product.name}
-          <br />
-          구매 수량: {totalQty}
-          <br />
-          발신자 이름: {data.sender}
-          <br />
-          메시지: {data.message || '(없음)'}
-        </div>,
-      );
-      goToHome();
-    } catch (err: any) {
-      if (err.message === '401') {
-        toast.error('로그인이 필요합니다.');
-        goToLogin();
-      } else {
-        toast.error(err.message);
-      }
-    }
-  };
-
-  const totalQty = receivers.reduce((sum, r) => sum + r.qty, 0);
-
   return (
     <MobileLayout>
       <Wrapper>
         <NavBar />
-
-        {/* 카드 선택 */}
-        <CardPicker
-          selectedId={selectedTemplate.id}
-          onSelect={(tpl) => {
-            setSelectedTemplate(tpl); // ✅ 선택한 템플릿 상태 업데이트
-            setValue('message', tpl.defaultTextMessage);
-          }}
-        />
-        {/* 카드 + 메세지 */}
-        <CardMessage
-          tpl={selectedTemplate}
-          register={register('message', {
-            validate: (v) => !isBlank(v) || '메시지를 입력해주세요.',
-          })}
-          error={errors.message?.message}
-        />
-
-        {/* 보내는 사람 */}
-        <SenderInfo
-          register={register('sender', {
-            validate: (v) => !isBlank(v) || '이름을 입력해주세요.',
-          })}
-          error={errors.sender?.message}
-        />
-
-        {/* 받는 사람 */}
-        <ReceiverInfo receivers={receivers} setReceivers={setReceivers} />
-
-        {/* 상품 정보 */}
-        <ProductInfo product={product} />
-
-        {/* 주문하기 버튼 */}
-        <OrderButton priceSum={product.price} qty={totalQty} onClick={handleSubmit(onSubmit)} />
+        <ErrorBoundary fallback={<div>상품 정보를 불러오는 중 오류가 발생했습니다.</div>}>
+          <Suspense
+            fallback={
+              <SpinnerWrapper>
+                <Spinner />
+              </SpinnerWrapper>
+            }
+          >
+            <OrderContent />
+          </Suspense>
+        </ErrorBoundary>
       </Wrapper>
     </MobileLayout>
   );
